Added kWh charge limit UI (non functional)
kWh charge limit will allow the user to specify a per-charge kWh limit.
This can be used to either not over-run a billing tier, or to make sure
to *just* charge enough to make it to a free charger for example.

Also added hourglass logo in preparation for delay timer/charge
scheduling.