Fix delayed event dispatching
Fixes a potential endless loop when a delayed event was posted from an
event handler. Since the cached "now" variable could be in the past
relative to the newly posted event time, delayed events could be fired
in a tight loop, causing a watchdog reset.

No current code causes this problem.