Remove relay self test; monitor relay output continuously
Removed the relay self test to make sure the J1772 connector does not
get energized when not connected to the car. Yes, this removes the
re-assuring "kachunk" sound when you start the unit.

Instead, the relay is now monitored continuously to make sure the output
of the sense pins always reflects the correct state of the relay and a
fault is triggered if it does not match.