Switch I2C bus to 400kHz
Massive potential for LCD glitching here...

The MCP23017 chip in the LCD backpack as well as the DS3231 clock chip
all support 400kHz I2C speed. However, interference can really be an
issue at this speed. Also, the value of the pull-up resistors is
important here to shape the edges of the pulses.

Initial testing is very positive with a massive increase in update
speed. The main loop is back to executing 36+ times a second and the LCD
is looking solid...